rename module from windchill to wc
